:sparkles: 半成品签到领积分兑物品 + 封号机制（有点烦，等不到开发完积分功能了）

+++ src/task/tasks/dofMaintenanceSql.ts
@@ -2,11 +2,26 @@ import config from "../../config/index.js"
 import { useKnex } from "../../db/knex.js"
 
 /**
+ * DOF 日常维护 SQL (10秒)
+ */
+export const dofMaintenanceSql10s = async () => {
+  // 创建封号信息 (因为不知道为什么会自动清理，每分钟干它一次，如果后面感觉不行，改成 3s 也不是不可以)
+  for (const banUser of config.banList) {
+    await useKnex().raw(`
+      INSERT IGNORE INTO d_taiwan.member_punish_info
+        (m_id, punish_type, occ_time, punish_value, apply_flag, start_time, end_time, admin_id, reason)
+      VALUES
+        (${banUser.dofUid}, 1, '2015-10-31 00:00:00', 101, 2, '2015-10-31 00:00:00', '9999-12-31 23:59:59', 'Cat73', 'GM')
+    `)
+  }
+}
+
+/**
  * DOF 日常维护 SQL (1分钟)
  */
 export const dofMaintenanceSql1m = async () => {
   // 清理封号信息
-  await useKnex().raw(`DELETE FROM d_taiwan.member_punish_info`)
+  await useKnex().raw(`DELETE FROM d_taiwan.member_punish_info WHERE admin_id = 'GM'`)
 }
 
 /**
